pyproject.toml
[project]
name = "fastzero"
version = "0.1.0"
description = ""
authors = [
    {name = "Gabriel Guedes",email = "gabrielguedesbsilva@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12, <4.0"
dependencies = [
    "fastapi (==0.111.0)",
    "sqlalchemy (>=2.0.36,<3.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "alembic (>=1.14.0,<2.0.0)",
    "pwdlib[argon2] (>=0.2.1,<0.3.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pytest-asyncio (>=0.25.2,<0.26.0)",
    "freezegun (>=1.5.1,<2.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.8.6"
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
taskipy = "^1.14.1"


[tool.poetry.group.deve.dependencies]
factory-boy = "^3.3.1"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check .; ruff check . --diff'
format = 'ruff check . --fix; ruff format .'
run = 'fastapi dev fastzero/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=fastzero -vv --showlocals --tb=long'
post_test = 'coverage html'

[tool.ruff.lint.pylint]
max-args = 10